+++ src/Application/Dx/SchemaCacheKeyGenerator.php
@@ -1,0 +1,17 @@
+<?php
+declare(strict_types=1);
+
+namespace Iriven\PhpFormGenerator\Application\Dx;
+
+/** @api */
+final class SchemaCacheKeyGenerator
+{
+    /**
+     * @param array<string, mixed> $context
+     */
+    public function generate(string $formName, array $context = []): string
+    {
+        ksort($context);
+        return sha1($formName . '|' . json_encode($context));
+    }
+}